handle "Exchange Server SE" - shares build number tree 15.2 with "Exchange Server 2019"
for whatever fucked up reason...

It's basically a "CU" with a completely new name.

Treat it as a different tree in internal representation, which means the
"root" trees for the main exchange version don't use the build number
prefix as primary key anymore; they might also need to check the version
name if the prefix isn't unique.